services/video_postprocess_service/requirements.txt
fastapi==0.115.12
pydantic==2.11.3
pydantic-settings==2.8.1
uvicorn==0.34.0
aio-pika==9.4.1
boto3==1.34.0
aiobotocore==2.13.2
aiofiles==23.2.1
pydantic==2.11.3
pydantic-settings==2.8.1
ffmpeg-python==0.2.0
orjson==3.10.16  